move around utility files for mse
