+++ robomimic/algo/algo.py
@@ -22,6 +22,7 @@ import robomimic.utils.tensor_utils as TensorUtils
 import robomimic.utils.torch_utils as TorchUtils
 import robomimic.utils.obs_utils as ObsUtils
 import robomimic.utils.action_utils as AcUtils
+import robomimic.utils.vis_utils as VisUtils
 
 from torch.utils.data import DataLoader
 
@@ -326,80 +327,6 @@ class Algo(object):
         """
         pass
 
-    def visualize(self, trainset, validset, savedir):
-        """
-        TODO: add documentation
-        """
-        pass
-
-    def make_model_prediction_plot(
-        self,
-        hdf5_path,
-        save_path,
-        images,
-        action_names,
-        actual_actions,
-        predicted_actions,
-    ):
-        """
-        TODO: documentation
-        actual_actions: (T, D)
-        predicted_actions: (T, D)
-        """
-        image_keys = sorted(list(images.keys()))
-        action_dim = actual_actions.shape[1]
-        traj_length = len(actual_actions)
-
-        # Plot
-        fig, axs = plt.subplots(len(images) + action_dim, 1, figsize=(30, (len(images) + action_dim) * 3))
-        for i, image_key in enumerate(image_keys):
-            interval = int(traj_length/15) # plot `5` images
-            images[image_key] = images[image_key][::interval]
-            combined_images = np.concatenate(images[image_key], axis=1)
-            axs[i].imshow(combined_images)
-            if i == 0:
-                axs[i].set_title(hdf5_path + '\n' + image_key, fontsize=30)
-            else:
-                axs[i].set_title(image_key, fontsize=30)
-            axs[i].axis("off")
-        for dim in range(action_dim):
-            ax = axs[len(images)+dim]
-            ax.plot(range(traj_length), actual_actions[:, dim], label='Actual Action', color='blue')
-            ax.plot(range(traj_length), predicted_actions[:, dim], label='Predicted Action', color='red')
-            # ax.set_xlabel('Timestep')
-            # ax.set_ylabel('Action Dimension {}'.format(dim + 1))
-            ax.set_title(action_names[dim], fontsize=30)
-            ax.xaxis.set_tick_params(labelsize=24)
-            ax.yaxis.set_tick_params(labelsize=24)
-            ax.legend(fontsize=20)
-        plt.subplots_adjust(left=0.05, right=0.95, top=0.95, bottom=0.05, wspace=0.3, hspace=0.6)
-        
-        # Save the figure with the specified path and filename
-        save_dir = os.path.dirname(save_path)
-        if not os.path.exists(save_dir):
-            os.makedirs(save_dir)
-        plt.savefig(save_path) 
-
-        fig.clear()
-        plt.close()
-        plt.cla()
-        plt.clf()
-
-    def get_action_names_for_vis(self, action_keys, training_sample):
-        """
-        TODO: documentation
-        """
-        modified_action_keys = [element.replace('action/', '') for element in action_keys]
-        action_names = []
-        
-        for i, action_key in enumerate(action_keys):
-            if isinstance(training_sample[action_key][0], np.ndarray):
-                action_names.extend([f'{modified_action_keys[i]}_{j+1}' for j in range(len(training_sample[action_key][0]))])
-            else:
-                action_names.append(modified_action_keys[i])
-
-        return action_names
-
 
 class PolicyAlgo(Algo):
     """
@@ -497,10 +424,17 @@ class PolicyAlgo(Algo):
         else:
             inference_datasets_mapping = {"Train": training_sampled_data}
 
-        action_names = self.get_action_names_for_vis(
-            action_keys=self.global_config.train.action_keys,
-            training_sample=training_sampled_data[0][0],
-        )
+        # extract action name for visualization
+        action_keys = self.global_config.train.action_keys
+        training_sample=training_sampled_data[0][0]
+        modified_action_keys = [element.replace("action/", "") for element in action_keys]
+        action_names = []
+        
+        for i, action_key in enumerate(action_keys):
+            if isinstance(training_sample[action_key][0], np.ndarray):
+                action_names.extend([f'{modified_action_keys[i]}_{j+1}' for j in range(len(training_sample[action_key][0]))])
+            else:
+                action_names.append(modified_action_keys[i])
 
         if visualize:
             print("Saving model prediction plots to {}".format(savedir))
@@ -518,7 +452,7 @@ class PolicyAlgo(Algo):
                 predicted_actions_all_traj.append(predicted_actions)
                 if visualize:
                     save_path = os.path.join(savedir, "{}_traj_{}.png".format(inference_key.lower(), traj_num))                
-                    self.make_model_prediction_plot(
+                    VisUtils.make_model_prediction_plot(
                         hdf5_path=d.hdf5_path,
                         save_path=save_path,
                         images=images,
@@ -545,7 +479,6 @@ class PolicyAlgo(Algo):
         
         return mse_log
     
-        
 
 class ValueAlgo(Algo):
     """
