+++ robomimic/scripts/config_gen/helper.py
@@ -859,8 +859,8 @@ def make_generator(args, make_generator_helper):
         time_str = datetime.datetime.fromtimestamp(time.time()).strftime('%m-%d-')
         args.name = time_str + str(args.name)
 
-    # if args.debug or args.tmplog:
-    #     args.no_wandb = True
+    if args.debug or args.tmplog:
+        args.no_wandb = True
 
     if args.wandb_proj_name is not None:
         # prepend data to wandb name
