+++ robomimic/utils/vis_utils.py
@@ -4,6 +4,7 @@ These functions can be a useful debugging tool.
 """
 import numpy as np
 import matplotlib.pyplot as plt
+import os
 
 import robomimic.utils.tensor_utils as TensorUtils
 import robomimic.utils.obs_utils as ObsUtils
@@ -90,3 +91,56 @@ def visualize_image_randomizer(original_image, randomized_image, randomizer_name
 
     # Show the entire grid of subplots
     plt.show()
+
+
+def make_model_prediction_plot(
+    hdf5_path,
+    save_path,
+    images,
+    action_names,
+    actual_actions,
+    predicted_actions,
+):
+    """
+    TODO: documentation
+    actual_actions: (T, D)
+    predicted_actions: (T, D)
+    """
+    image_keys = sorted(list(images.keys()))
+    action_dim = actual_actions.shape[1]
+    traj_length = len(actual_actions)
+
+    # Plot
+    fig, axs = plt.subplots(len(images) + action_dim, 1, figsize=(30, (len(images) + action_dim) * 3))
+    for i, image_key in enumerate(image_keys):
+        interval = int(traj_length/15) # plot `5` images
+        images[image_key] = images[image_key][::interval]
+        combined_images = np.concatenate(images[image_key], axis=1)
+        axs[i].imshow(combined_images)
+        if i == 0:
+            axs[i].set_title(hdf5_path + '\n' + image_key, fontsize=30)
+        else:
+            axs[i].set_title(image_key, fontsize=30)
+        axs[i].axis("off")
+    for dim in range(action_dim):
+        ax = axs[len(images)+dim]
+        ax.plot(range(traj_length), actual_actions[:, dim], label='Actual Action', color='blue')
+        ax.plot(range(traj_length), predicted_actions[:, dim], label='Predicted Action', color='red')
+        # ax.set_xlabel('Timestep')
+        # ax.set_ylabel('Action Dimension {}'.format(dim + 1))
+        ax.set_title(action_names[dim], fontsize=30)
+        ax.xaxis.set_tick_params(labelsize=24)
+        ax.yaxis.set_tick_params(labelsize=24)
+        ax.legend(fontsize=20)
+    plt.subplots_adjust(left=0.05, right=0.95, top=0.95, bottom=0.05, wspace=0.3, hspace=0.6)
+    
+    # Save the figure with the specified path and filename
+    save_dir = os.path.dirname(save_path)
+    if not os.path.exists(save_dir):
+        os.makedirs(save_dir)
+    plt.savefig(save_path) 
+
+    fig.clear()
+    plt.close()
+    plt.cla()
+    plt.clf()
